import of CINT 5.15.103. For what else is new see http://root.cern.ch/root/Cint.phtml?relnote. Removed C++ comments from shl.c.
git-svn-id: http://root.cern.ch/svn/root/trunk@7312 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/src/v6_quote.cxx
@@ -601,7 +601,7 @@ char *result;
 	flag=0;
 	ii=0;
 	for(i=0;i<32;i++) {
-		if((32-i)<=dig) flag=1;
+		if((int)(32-i)<=dig) flag=1;
 		switch(result[i]){
 		case '0':
 			if(tristate[i]=='0') {
